Primary implementation for list item support
See #10.

Outstanding issues
==================

Nothing will happen if you check/uncheck the box in the rendered
HTML. Need to attach some JavaScript to get the click event and
update the source accordingly. To achieve this we need a way to
map checkboxes so that we can find the source for it. ID tags may
be messy, but I don't see another way. Maybe we should just disable
the checkboxes, fuck it.

Will need to provide extra stylesheets to hide the list item's
bullet/number when a task list is detected. This should be global
I think---Styles can override this if they want to.

Need to add an extra prefernce for it, and some extra code to load
it into the renderer. We should be able to reuse the same renderer,
just need to change the flag.

The preference switch should be put in the HTML pane, I think.

Need to add hoextdown license and reference.