Disable underscore emph when MathJax is on
Implemented by always turning underline extension on (which is the
ONLY way to distinguish _this_ and *this*) when MathJax is on, and
use a custom renderer callback with an extra HTML_* flag.

This is pretty hacky, but at least it works. Is it possible to
convice Hoedown to add functionality in substituting parsing
callbacks? Might worth a try, need to choose the word, though. :)

Fix #8