Fill split view background so avoid window glitch
When scrolling too fast the left scroll view might be a little
smaller than the split view's immediate content view (which is
a custom NSView) due to syntax highlighter-triggered resizing,
resulting a gutter at the bottom of the document. Worked around
by painting the split view with editor background (so that the
gutter is not visible).