Setup edito only when editor settings change
Fixes flickers on preferences change (unless it's an editor
preference). Fix #123.

This can be furthur refined by re-initialising only NEEDED parts
based on exactly WHAT changed. Maybe we should change the signature
of -setupEditor to -setupEditor: to take an additional NSString
as the name of preference changed. Passed nil to initialise
everything. Will work on this later.