Implement better file-opening utility
This implements a preference-based file-opening scheme. When the
CLI utility is run, all file name arguments are resolved to
absolute URLs and stored in a user default key *under the master
Macdown application's namespace*, and then launch MacDown.

Everytime MacDown is launched, it checks the aforementioned user
default key, and open any documents in it, creating non-saved drafts
if necessary (for non-existing documents). This mimics the expected
behaviours for "normal" editors, like VIM, nano, etc.. The user
default key is reset to nil after the documents are opened.

Closed #35.