Adjust insets when editor size changes
This fixes a problem in which editor width limits are not correctly
applied if the window changes size. This happens because the width
limit is actually a hack that depends on container insets. If insets
don't change when the editor resizes, the editor's width (i.e.
total width - 2 * insets) will not be correct.