Insert blockquote prefix for blank lines too
Similar logic to blank-line indentation. See 61e4a73 and 4d98888.

Fix #383.

Side effect: This patch also make list item insertion work with
empty lines - a backward-incompatible behaviAoural change. But this
seems to be what people want, so let's roll with it for now. We can
always change this later if reports appear.