Upgrade CocoaPods and Sparkle
This addresses the important security fix in Sparkle 1.13.1.
See <https://github.com/sparkle-project/Sparkle/releases/tag/1.13.1>.

Thankfully CocoaPods now fixes the ARC issue blocked by Handlebars-objc,
so that we can upgrade to 0.39.0, which is compatible with new versions
of Sparkle.

Handlebars-objc has also upgraded to 1.4.4 so that it works with
newer CocoaPods versions.