Provide better error for nonexistent link target
UX enhancement regarding #580.

Previously MacDown delegates error-handling to Finder when a link
to nonexistent local file is clicked in preview (via NSWorkspace).
This leads to a generic "Can't find the file XXX" message, which
is not really useful.

This patch detects a file URL in the above scenario, and shows a
better message inside MacDown that provides an absolute path
(instead of just the filename), and hints the user to use the
auto-create feature if desired.

Behaviours are unchanged for cases which...

* The URL is not a local file.
* The target file exists.
* The auto-create option is on.