Dynamically replace content instead of reloading
This works around the WebView caching problem. By reusing the same DOM
as much as possible, we can avoid reloading external resources again
and again every time we render. This should improve performance on
large documents.

Note that the current implementation is quite crude. We always get the
whole HTML document, and extract the part we need after putting the
whole thing together (using regex). Ideally, since the style and script
section never updates during document editing, and the main content
never updates on pref changes, we should be able to only notify changes
for needed parts seperately. But we lack that kind of renderer
notification. This would be a target for refactoring in the future.