Refactor file-launching method in cmd
Previously the cmd util uses different strategies to open files based
on whether MacDown is currently running. This results in various
problems, the most prominent being that NSWorkspace cannot be used
to pass a path for files that do not exist.

This change unifies the opening strategy to always be based on
cross-application NSUserDefaults instead, and moves the open file
check from during application launch to application activation.
This makes the implementation simpler, and also solves the problem
aforementioned.

Fix #672.