Refactor UI-controller sync into setter
This should make the logic code a lot cleaner, with the expense of a
very fat property setter. This should probably be further refactored
into data binding or KVO, but that would be a performance hit. For
one single property this is readable enough.

Also changed many setter/getter usage into dot syntax.