UI for CLI utility installation
Implemented logic for CLI installation from the GUI app.

A preference pane is added. The pane detects whether the user has
something installed that looks like the MacDown CLI utility by checking
common installtion locations, e.g. Homebrew.

If the util is not detected to be installed, the user can click a
button to install it. For now the only option is to install it to
/usr/local/bin/macdown.