Added GitHub Tomorrow Style
"GitHub Tomorrow" is a preview-pane theme that combines the basic formatting of the GitHub preview-pane theme with the colors of the Tomorrow edit-pane theme. I created this because there was no dark-mode preview-pane theme that looked nice with the Tomorrow+ edit-pane theme. The preview-pane background color is slightly lighter than the edit-pane background color in order to provide contrast with Tomorrow-themed inline code blocks.